Implement HW02 - the main part

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,46 +1,78 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
+    private final Logger log = LoggerFactory.getLogger(getClass());
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
-    {
-        MealsUtil.MEALS.forEach(this::save);
-    }
-
     @Override
     public Meal save(Meal meal) {
+
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
         // treat case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        Meal result = repository.computeIfPresent(meal.getId(),
+                (id, oldMeal) -> Objects.equals(oldMeal.getUserId(), meal.getUserId()) ? meal : oldMeal);
+
+        //this check is needed because, if return null is inside computeIfPresent - the key is removed
+        return Objects.equals(result, meal) ? result : null;
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, int userID) {
+        Meal meal = get(id, userID);
+
+        if (meal == null) {
+            return false;
+        }
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userID) {
+        log.debug("Get id {}, userID {}", id, userID);
+        Meal meal = repository.get(id);
+
+        if (meal == null) {
+            return null;
+        } else if (meal.getUserId() != userID) {
+            return null;
+        }
+
+        return meal;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public List<Meal> getAll() {
+        log.debug("getAll items {} , size {}", repository, repository.size());
+        return repository.values().stream()
+                .sorted((meal1, meal2) -> meal2.getDateTime().compareTo(meal1.getDateTime()))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<Meal> getAllByUserId(int userID) {
+        log.debug("getAll id {} items {}", userID, repository.values());
+        return repository.values().stream()
+                .filter(meal -> (meal.getUserId() == userID))
+                .sorted((meal1, meal2) -> meal2.getDateTime().compareTo(meal1.getDateTime()))
+                .collect(Collectors.toList());
     }
 }
